+++ src/client_core.d
@@ -1,0 +1,2 @@
+src/client_core.o: src/client_core.c src/client_core.h
+src/client_core.h: