+++ src/main.d
@@ -10,8 +10,9 @@ src/main.o: src/main.c vendor/ffmpeg/include/libavutil/avutil.h \
  vendor/ffmpeg/include/libavutil/mathematics.h \
  vendor/ffmpeg/include/libavutil/intfloat.h \
  vendor/ffmpeg/include/libavutil/log.h \
- vendor/ffmpeg/include/libavutil/pixfmt.h src/player.h src/plugin.h \
- src/plugin_loader.h src/lang.h src/update.h src/config.h src/cd.h
+ vendor/ffmpeg/include/libavutil/pixfmt.h src/player.h src/client_core.h \
+ src/stream_player.h src/plugin.h src/plugin_loader.h src/lang.h \
+ src/update.h src/config.h src/cd.h
 vendor/ffmpeg/include/libavutil/avutil.h:
 vendor/ffmpeg/include/libavutil/common.h:
 vendor/ffmpeg/include/libavutil/attributes.h:
@@ -26,6 +27,8 @@ vendor/ffmpeg/include/libavutil/intfloat.h:
 vendor/ffmpeg/include/libavutil/log.h:
 vendor/ffmpeg/include/libavutil/pixfmt.h:
 src/player.h:
+src/client_core.h:
+src/stream_player.h:
 src/plugin.h:
 src/plugin_loader.h:
 src/lang.h: