+++ src/stream_player.d
@@ -1,0 +1,9 @@
+src/stream_player.o: src/stream_player.c src/../vendor/miniaudio.h \
+ src/stream_player.h src/player.h src/plugin_loader.h src/plugin.h \
+ src/client_core.h
+src/../vendor/miniaudio.h:
+src/stream_player.h:
+src/player.h:
+src/plugin_loader.h:
+src/plugin.h:
+src/client_core.h: